requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
google-cloud-bigquery==3.11.4
python-dotenv==1.0.0
pydantic==2.4.2 